inclusion of modified benchmark functions 'benchmark_1d_potential_subplots' and 'benchmark_1d_potential_morse'.
The function 'benchmark_1d_potential_subplots' now enables a more comparative visualization of the effect of varying parameters on the general shape of the tested potential, force, and Hessian. The function 'benchmark_1d_potential_morse' is an adapted variant of the 'benchmark_1d_potential' function that added some Morse potential-specific visualization features and enables the user to save the generated figures as PDFs in the subdirectory 'Figures'.